+++ src/ui/components/tests/layer-conversation-view-test.js
@@ -344,12 +344,12 @@ describe('layer-conversation-view', function() {
     });
   });
 
-  describe("The getMenuOptions property", function() {
-    it("Should set the list getMenuOptions property", function() {
+  describe("The getMenuItems property", function() {
+    it("Should set the list getMenuItems property", function() {
       var f = function() {};
-      el.getMenuOptions = f;
+      el.getMenuItems = f;
       Layer.Utils.defer.flush();
-      expect(el.nodes.list.getMenuOptions).toBe(f);
+      expect(el.nodes.list.getMenuItems).toBe(f);
     });
   });